R171219004: added substraction function

+++ substract.cpp
@@ -1,12 +1,12 @@
 #include<iostream>
 #include<cmaths>
 using namespace std;
-int substract(int x,int y);
+float substract(float x,float y);
 int main()
 {
-int first_no;
-int second_no;
-int diff_of_nos;
+float first_no;
+float second_no;
+float diff_of_nos;
 
 cout<<"enter the 2 numbers"<<endl;
 cin>>first_no>>second_no;
@@ -14,7 +14,7 @@ diff_of_nos= substract(first_no, second_no);
 cout<<"difference of two numbers"<<diff_of_nos<<endl;
 return 0;
 }
-int substract(int x,int y)
+float substract(float x,float y)
 {
 if(x>y)
 	return (x-y);
